Moved missing octavia pool-member fields to package loadbalancer from networking extensions
