+++ openstack/loadbalancer/v2/pools/results.go
@@ -1,6 +1,9 @@
 package pools
 
 import (
+	"encoding/json"
+	"time"
+
 	"github.com/gophercloud/gophercloud"
 	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/monitors"
 	"github.com/gophercloud/gophercloud/pagination"
@@ -204,6 +207,24 @@ type Member struct {
 	// The provisioning status of the pool.
 	// This value is ACTIVE, PENDING_* or ERROR.
 	ProvisioningStatus string `json:"provisioning_status"`
+
+	// DateTime when the member was created
+	CreatedAt time.Time `json:"-"`
+
+	// DateTime when the member was updated
+	UpdatedAt time.Time `json:"-"`
+
+	// The operating status of the member
+	OperatingStatus string `json:"operating_status"`
+
+	// Is the member a backup? Backup members only receive traffic when all non-backup members are down.
+	Backup bool `json:"backup"`
+
+	// An alternate IP address used for health monitoring a backend member.
+	MonitorAddress string `json:"monitor_address"`
+
+	// An alternate protocol port used for health monitoring a backend member.
+	MonitorPort int `json:"monitor_port"`
 }
 
 // MemberPage is the page returned by a pager when traversing over a
@@ -245,6 +266,23 @@ func ExtractMembers(r pagination.Page) ([]Member, error) {
 
 type commonMemberResult struct {
 	gophercloud.Result
+}
+
+func (r *Member) UnmarshalJSON(b []byte) error {
+	type tmp Member
+	var s struct {
+		tmp
+		CreatedAt gophercloud.JSONRFC3339NoZ `json:"created_at"`
+		UpdatedAt gophercloud.JSONRFC3339NoZ `json:"updated_at"`
+	}
+	err := json.Unmarshal(b, &s)
+	if err != nil {
+		return err
+	}
+	*r = Member(s.tmp)
+	r.CreatedAt = time.Time(s.CreatedAt)
+	r.UpdatedAt = time.Time(s.UpdatedAt)
+	return nil
 }
 
 // ExtractMember is a function that accepts a result and extracts a member.
